+++ src/NullKill/NullKill.fs
@@ -1,30 +1,29 @@
 module NullKill
 
 open System.Reflection
-open System.Collections.Generic
 open Microsoft.FSharp.Reflection
 
-let rec CheckEnumerable thing =
+let rec private CheckEnumerable depth thing =
     let e = box thing :?> System.Collections.IEnumerable
     let typed =
         e
         |> Seq.cast
     if Seq.length typed > 0 then
         typed
-        |> Seq.map (fun i -> HasNoNulls' i <| i.GetType())
+        |> Seq.map (fun i -> HasNoNulls' (depth + 1) i <| i.GetType())
         |> Seq.reduce (&&)
     else
         true
 
-and CheckField thing (field : FieldInfo) =
+and private CheckField depth thing (field : FieldInfo) =
     if field.FieldType.IsValueType then
         true
     else
         let o = field.GetValue(thing)
         let t = field.FieldType
-        HasNoNulls' o t
+        HasNoNulls' (depth + 1) o t
         
-and CheckFields thing (t : System.Type) =
+and private CheckFields depth thing (t : System.Type) =
     if t.IsValueType then
         true
     else
@@ -33,10 +32,10 @@ and CheckFields thing (t : System.Type) =
             true
         | fields ->
             fields
-            |> Array.map (CheckField thing)
+            |> Array.map (CheckField depth thing)
             |> Array.reduce (&&)
 
-and CheckProperty thing (prop : PropertyInfo) =
+and private CheckProperty depth thing (prop : PropertyInfo) =
     if prop.PropertyType.IsValueType || prop.GetIndexParameters().Length > 0 then
         true
     else
@@ -45,11 +44,11 @@ and CheckProperty thing (prop : PropertyInfo) =
         let o = prop.GetValue(thing, null)
         let t = prop.PropertyType
         if enumerables.Length > 0 then
-            HasNoNulls' o t && (CheckEnumerable o)
+            HasNoNulls' (depth + 1) o t && (CheckEnumerable depth o)
         else
-            HasNoNulls' o t
+            HasNoNulls' (depth + 1) o t
 
-and CheckProperties thing (t : System.Type) =
+and private CheckProperties depth thing (t : System.Type) =
     if t.IsValueType then
         true
     else
@@ -58,20 +57,27 @@ and CheckProperties thing (t : System.Type) =
             true
         | fields ->
             fields
-            |> Array.map (CheckProperty thing)
+            |> Array.map (CheckProperty depth thing)
             |> Array.reduce (&&)
 
-and private HasNoNulls' (thing : obj) thingType =
-    if thing = null then
+and private HasNoNulls' depth (thing : obj) thingType =
+    match thing with
+    | _ when depth > 50 ->
+        true
+    | t when (box t) = null ->
         if FSharpType.IsUnion thingType then
             true
         else
             false
-    else
-        CheckProperties thing thingType && CheckFields thing thingType
+    | _ ->
+        match thingType with
+        | t when t = typeof<string> ->
+            true
+        | _ ->
+            CheckProperties depth thing thingType && CheckFields depth thing thingType
 
 let HasNoNulls<'a> (thing : 'a) =
-    HasNoNulls' thing (typedefof<'a>)
+    HasNoNulls' 0 thing (typedefof<'a>)
 
 let Check thing =
     if HasNoNulls thing then